+++ src/Application/Communication/AudienceMode.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EventFlow\Application\Communication;
+
+enum AudienceMode: string
+{
+    case DYNAMIC = 'dynamic';
+    case SNAPSHOT = 'snapshot';
+}