+++ src/Application/Communication/CampaignPurpose.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace EventFlow\Application\Communication;
+
+enum CampaignPurpose: string
+{
+    case INVITATION = 'invitation';
+    case REMINDER = 'reminder';
+    case EVENT_UPDATE = 'event_update';
+    case OPERATIONAL = 'operational';
+}