+++ src/Application/Communication/CommunicationChannel.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EventFlow\Application\Communication;
+
+enum CommunicationChannel: string
+{
+    case EMAIL = 'email';
+    case SMS = 'sms';
+}